requirements.txt
#requirements.txt
# 原有依赖
neo4j==5.14.1
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
ollama==0.5.3
psutil==5.9.6

# GraphRAG新增依赖
chromadb==0.4.15
numpy==1.24.3
scikit-learn==1.3.0

# 备用嵌入模型依赖（当远程Ollama不可用时）
sentence-transformers==2.2.2

# 注意：主要使用远程Ollama服务提供bge-m3:latest嵌入模型
# sentence-transformers仅作为连接失败时的备用方案